requirements.txt
pandas
matplotlib
seaborn
fpdf2
